requirements.txt
torch>=2.0.0
torchvision>=0.15.0
segmentation-models-pytorch>=0.3.3
albumentations==1.3.0
numpy>=1.24.0
opencv-python>=4.7.0
Pillow>=9.4.0
scikit-learn>=1.2.0
scipy>=1.10.0
matplotlib>=3.7.0
tifffile>=2023.1.23
tqdm>=4.65.0
ipywidgets>=8.0.0
onnx>=1.14.0
onnxruntime>=1.15.0
jupyter>=1.0.0
